# ENH: Optimize np.power for integer type