--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -476,7 +476,11 @@ NPY_NO_EXPORT void
 {
     if (steps[1]==0) {
         // stride for second argument is 0
-        BINARY_DEFS
+        char *ip1 = args[0], *ip2 = args[1], *op1 = args[2];
+        npy_intp is1 = steps[0], os1 = steps[2];
+        npy_intp n = dimensions[0];
+        npy_intp i;
+
         @type@ in2 = *(@type@ *)ip2;
         #if @SIGNED@
                 if (in2 < 0) {