--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -471,7 +471,8 @@ NPY_NO_EXPORT NPY_GCC_OPT_3 int
 }
 /**end repeat1**/
 
-static inline @type@ _@TYPE@_squared_exponentiation_helper(@type@ base, @type@ exponent_two, int first_bit) {
+static inline @type@
+_@TYPE@_squared_exponentiation_helper(@type@ base, @type@ exponent_two, int first_bit) {
    // Helper method to calculate power using squared exponentiation
    // The algorithm is partly unrolled. The second and third argument are the exponent//2 and the first bit of the exponent
    @type@ out = first_bit ? base : 1;
@@ -493,11 +494,11 @@ NPY_NO_EXPORT void
         BINARY_DEFS_ZERO_STRIDE
         const @type@ in2 = *(@type@ *)ip2;
         #if @SIGNED@
-                if (in2 < 0) {
-                    npy_gil_error(PyExc_ValueError,
-                                  "Integers to negative integer powers are not allowed.");
-                    return;
-                }
+            if (in2 < 0) {
+                npy_gil_error(PyExc_ValueError,
+                              "Integers to negative integer powers are not allowed.");
+                return;
+            }
         #endif
 
         int first_bit = in2 & 1;