--- a/numpy/_core/src/umath/loops.c.src
+++ b/numpy/_core/src/umath/loops.c.src
@@ -491,7 +491,7 @@ NPY_NO_EXPORT void
 {
     if (steps[1]==0) {
         // stride for second argument is 0
-        BINARY_DEFS_ZERO_STRIDE
+        BINARY_DEFS
         const @type@ in2 = *(@type@ *)ip2;
         #if @SIGNED@
             if (in2 < 0) {
@@ -504,7 +504,7 @@ NPY_NO_EXPORT void
         int first_bit = in2 & 1;
         @type@ in2start = in2 >> 1;
 
-        BINARY_LOOP_SLIDING_ZERO_STRIDE {
+        BINARY_LOOP_SLIDING {
             @type@ in1 = *(@type@ *)ip1;
 
             *((@type@ *) op1) = _@TYPE@_squared_exponentiation_helper(in1, in2start, first_bit);