In this PR we optimize `np.power(x, n)` for integer types and a scalar argument `n`. The current implementation is a generic binary loop for the arguments. In the case `n` is a scalar (stride 0) we can optimize the  loop. 

Micro benchmark:
```
import numpy as np
import timeit

x=np.arange(100_000)
n = 20_000

dt=timeit.timeit('x**2', globals=globals(), number=n)
print(f'x**2 {1e6*dt/n}')
dt=timeit.timeit('x**3', globals=globals(), number=n)
print(f'x**3 {1e6*dt/n}')
dt=timeit.timeit('np.power(x, 1)', globals=globals(), number=n)
print(f'power(x, 1) {1e6*dt/n}')
dt=timeit.timeit('np.power(x, 2)', globals=globals(), number=n)
print(f'power(x, 2) {1e6*dt/n}')
dt=timeit.timeit('np.power(x, 3)', globals=globals(), number=n)
print(f'power(x, 3) {1e6*dt/n}')
dt=timeit.timeit('np.power(x, 4)', globals=globals(), number=n)
print(f'power(x, 4) {1e6*dt/n}')
```
Main:
```
x**2 62.292819999856874
x**3 135.96379000227898
power(x, 1) 103.64716000040062
power(x, 2) 134.2153200006578
power(x, 3) 142.9221049998887
power(x, 4) 197.58593500009738
```
PR:
```
x**2 59.52026000013575
x**3 87.76193000085186
power(x, 1) 62.03027000010479
power(x, 2) 86.0 (estimated)
power(x, 3) 86.40300999977626
power(x, 4) 126.56257500057109

```
Notes

* We could add the special case `if (in2==2)` to improve performance. For `np.power(x, 2)` we have per loop: main 143.2 us, special case exponent 2 55.7 us, no special case about 86 us.
* Related: #9363, #6399, #22651, #7786