@eendebakpt 2024-03-21
@ngoldbaum This PR is similar to #26055. Since we are dealing with integers there are no (potential) issues with accuracy and we can optimize for all exponents. On the other hand: the gain is smaller and I am not so sure about the use case for integers power.