@ngoldbaum 2024-04-17
It would be nice to see a proper asv benchmark for this.