@eendebakpt 2024-04-17
> It would be nice to see a proper asv benchmark for this.

I will try to do this (might take some time though, the spin command is not functioning well on my system)