@eendebakpt 2024-04-27
> It would be nice to see a proper asv benchmark for this.

I added a few benchmarks for the integer `np.power` case. Results on main:
```
· Running 3 total benchmarks (1 commits * 1 environments * 3 benchmarks)
[ 0.00%] ·· Benchmarking existing-pyC__projects_misc_numpy_.venv_Scripts_python.exe
[16.67%] ··· Running (bench_ufunc.BinaryBenchInteger.time_pow--)...
[66.67%] ··· bench_ufunc.BinaryBenchInteger.time_pow                                                                                                                              ok
[66.67%] ··· ============= =============
                 dtype
             ------------- -------------
              numpy.int32    5.16±0.3ms
              numpy.int64   5.52±0.05ms
             ============= =============

[83.33%] ··· bench_ufunc.BinaryBenchInteger.time_pow_five                                                                                                                         ok
[83.33%] ··· ============= =============
                 dtype
             ------------- -------------
              numpy.int32   2.32±0.05ms
              numpy.int64    3.05±0.2ms
             ============= =============

[100.00%] ··· bench_ufunc.BinaryBenchInteger.time_pow_two                                                                                                                          ok
[100.00%] ··· ============= =============
                  dtype
              ------------- -------------
               numpy.int32    2.02±0.2ms
               numpy.int64   2.76±0.07ms
              ============= =============

```
Results for this PR:
```
· Running 3 total benchmarks (1 commits * 1 environments * 3 benchmarks)
[ 0.00%] ·· Benchmarking existing-pyC__projects_misc_numpy_.venv_Scripts_python.exe
[16.67%] ··· Running (bench_ufunc.BinaryBenchInteger.time_pow--)...
[66.67%] ··· bench_ufunc.BinaryBenchInteger.time_pow                                                                ok
[66.67%] ··· ============= ============
                 dtype
             ------------- ------------
              numpy.int32   5.07±0.3ms
              numpy.int64   5.89±0.3ms
             ============= ============

[83.33%] ··· bench_ufunc.BinaryBenchInteger.time_pow_five                                                           ok
[83.33%] ··· ============= =============
                 dtype
             ------------- -------------
              numpy.int32   1.54±0.04ms
              numpy.int64   2.24±0.09ms
             ============= =============

[100.00%] ··· bench_ufunc.BinaryBenchInteger.time_pow_two                                                            ok
[100.00%] ··· ============= =============
                  dtype
              ------------- -------------
               numpy.int32   1.17±0.06ms
               numpy.int64   1.97±0.08ms
              ============= =============
```