@ngoldbaum 2024-05-06
On the latest version of this PR, on my laptop:

```
| Change   | Before [dd33199e] <main>   | After [561348b4] <faster_integer_power>   |   Ratio | Benchmark (Parameter)                                               |
|----------|----------------------------|-------------------------------------------|---------|---------------------------------------------------------------------|
| -        | 1.77±0.02ms                | 1.19±0.02ms                               |    0.68 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int64'>) |
| -        | 1.50±0.06ms                | 964±40μs                                  |    0.64 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int64'>)  |
| -        | 1.29±0.03ms                | 747±10μs                                  |    0.58 | bench_ufunc.BinaryBenchInteger.time_pow_two(<class 'numpy.int32'>)  |
| -        | 2.36±0.03ms                | 995±20μs                                  |    0.42 | bench_ufunc.BinaryBenchInteger.time_pow_five(<class 'numpy.int32'>) |
```

And no slowdowns. @seberg do the low-level code changes in the final version of this look good to you?